Fix binary bloat when using local static vars
In case static local variables are used anywhere in the C++ code of
the sketch or the libraries, gcc kicks in to protect initializing
those from any potential threads being run in parallel.

The implementation of the above is to emit calls to __cxa_guard_acquire()
and __cxa_guard_release(). These functions are part of libstdc++ [1],
and so, once a local static variable appears, the whole libstdc++ is
linked into the binary with all its bells and whistles, greatly
increasing the binary size as a result.

One possible solution to that was proposed in [2], but it dealt with
a single consequence of linking in libstdc++ (adding the
__verbose_terminate_handler()).

This commits deals with the reason of libstdc++ linking. A gcc option
-fno-threadsafe-statics is disabling the above local statics protection
mechanism, as a result libstdc++ is not linked in.

Resulting binary size comparison using a sketch from FastLED examples
(as reported by Arduino IDE, in bytes; the first number is "program
storage space", the second is "dynamic memory"):

Before: 75160 / 6392
After:  14964 / 3152

For comparison, here are the numbers after applying the commit from [2]:

PR#28:	29592 / 6328

Finally, the resulting sketch is working just fine.

[1] https://github.com/gcc-mirror/gcc/blob/master/libstdc%2B%2B-v3/libsupc%2B%2B/guard.cc
[2] https://github.com/rogerclarkmelbourne/Arduino_STM32/pull/28

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>